feat(api,core,database): implement user CRUD use cases
Implement the stub methods in UserUseCasesImpl with full database integration:
- add_user: creates a new user within a transaction
- update_user: updates an existing user with optimistic locking
- delete_user: removes a user and returns the deleted record
- find_by_id: retrieves a user by ID using read-only transaction
- list_users: paginated user listing with cursor-based navigation

Add delete_user and list_users to the UserService port trait and implement
them in the database adapter. list_users defaults to 50 results and caps
page_size at 50.

Add input validation with new error types:
- InvalidId: rejects negative IDs
- InvalidPageSize: rejects page_size < 1

Update API error handling to return 400 Bad Request for validation errors.

Pair-programmed with Claude Code - https://claude.com/claude-code

+++ crates/api/src/http/user.rs
@@ -61,8 +61,8 @@ pub(crate) mod routes {
 
     #[derive(Deserialize, Debug, Default)]
     pub struct FilterOptions {
-        pub start_id: Option<usize>,
-        pub page_size: Option<usize>,
+        pub start_id: Option<i64>,
+        pub page_size: Option<u64>,
     }
 
     #[derive(Serialize, Debug)]
